forgot to compile and needed to change i to an int

+++ search_algorithms/1-binary.c
@@ -8,7 +8,7 @@
 */
 int binary_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	int i = 0;
 	int min = 0, max = size - 1, mid = 0;
 
 	if (!array || size == 0)
